BREAKING CHANGE(cleanup): Convert to TypeScript
v5.0.0-alpha.1

We convert `lock_and_cache` to TypeScript, and we apply `eslint` and
`prettier` with default arguments.

This patch tries to touch as little of the original logic as
possible. We don't convert any more code to async, and we only remove
a small amount of possibly-unused code that we couldn't make
typecheck.

We also fix a broken test case that used `map`, and accidentally
passed too many arguments to the wrapped function.