v6.0.0-alpha.1: Extensively overhaul
This PR ports the code to TypeScript and rewrites most of it, stripping
out `cache-manager` because it's hard to debug and we only need
something simple.

The new API is fairly similar to the old, but many minor utility
functions have been removed and the constructor is different.

We forward-ported the excellent 4.x tests and they all pass (using the
new API).